+++ packages/app/src/cli/commands/app/dev.ts
@@ -1,15 +1,11 @@
 import {appFlags} from '../../flags.js'
-import dev from '../../services/dev.js'
-import {dev as dev2} from '../../services/dev2.js'
+import {dev} from '../../services/dev.js'
 import Command from '../../utilities/app-command.js'
 import {showApiKeyDeprecationWarning} from '../../prompts/deprecation-warnings.js'
 import {Flags} from '@oclif/core'
 import {normalizeStoreFqdn} from '@shopify/cli-kit/node/context/fqdn'
 import {globalFlags} from '@shopify/cli-kit/node/cli'
 import {addPublicMetadata} from '@shopify/cli-kit/node/metadata'
-import {isTruthy} from '@shopify/cli-kit/node/context/utilities'
-import {getEnvironmentVariables} from '@shopify/cli-kit/node/environment'
-import {isShopify} from '@shopify/cli-kit/node/context/local'
 
 export default class Dev extends Command {
   static description = 'Run the app.'
@@ -134,12 +130,6 @@ export default class Dev extends Command {
       notify: flags.notify,
     }
 
-    const enableNewDev = getEnvironmentVariables().SHOPIFY_CLI_NEW_DEV
-
-    if (isTruthy(enableNewDev) || (enableNewDev === undefined && (await isShopify()))) {
-      await dev2(devOptions)
-    } else {
-      await dev(devOptions)
-    }
+    await dev(devOptions)
   }
 }
